% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.data}
\alias{generate.data}
\title{Generate simulated data}
\usage{
generate.data(p, n, K, network.type, umin, umax)
}
\arguments{
\item{p}{The number of genes.}

\item{n}{The sample size. A positive intege or a vector with length equal to the number of cancer subtypes.}

\item{K}{The number of cancer subtypes.}

\item{network.type}{A character string indicating which network type is generated. "ER" (Erdos-Renyi) and "SF" (scale-free) can be used.}

\item{umin}{The lower limits of the edge values.}

\item{umax}{The upper limits of the edge values.}
}
\value{
\item{\code{X}}{A list (length = \eqn{K}) of data matrices (\eqn{n_k \times p}), where K is the number of cancer subtypes and \eqn{n_k} is the sample size of k-th cancer subtype.}
\item{\code{theta.y}}{ A list (length = \eqn{K}) of precision matrices (\eqn{p \times p}) from cancerous cells of different  cancer subtyes, where K is the number of cancer subtypes
                      and \eqn{n_k} is the sample size of of k-th cancer subtype.}
\item{\code{theta.z}}{ A matrix of precision matrice  (\eqn{p \times p}) from non-cancerous cells shared by all cancer subtyes.}
\item{\code{purity}}{A list (length = \eqn{K}) of the tumor purity information vectors (\eqn{n_k \times 1}), where K is the number of cancer subtypes
and \eqn{n_k} is the sample size of of k-th cancer subtype.}
}
\description{
The complete procedure for generating simulated data. For details, refer to simulation study (Section 3.1 in the main text).
}
\details{
The function is used to generate the gene expression datasets.
}
\examples{
# Simulation data
data.x = generate.data(p = 100, n = 100, K = 4, network.type = "ER", umin = 0.5, umax = 1)
}
\references{
Jia-Juan Tu, Le Ou-Yang, Hong Yan, Xiao-Fei Zhang and Hong Qin (2019), Joint reconstruction of multiple gene networks by simultaneously capturing inter-tumor
     and intra-tumor heterogeneity.
}
\seealso{
{ \code{\link{NETI2}}, \code{\link{TCGA.BRCA}}}
}
\author{
Jia-Juan Tu
}
