% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NETI2.R
\name{NETI2}
\alias{NETI2}
\title{Joint reconstruction of multiple gene NETworks by simultaneously capturing Inter-tumor and Intra-tumor heterogeneity}
\usage{
NETI2(X,purity,lambda, tau, delta)
}
\arguments{
\item{X}{A list (length = \eqn{K}) of data matrices (\eqn{n_k \times p}), where K is the number of cancer subtypes and \eqn{n_k} is the sample size of k-th cancer subtype.}

\item{purity}{A list (length = \eqn{K}) of the tumor purity information vectors (\eqn{n_k \times 1}), where K is the number of cancer subtypes
and \eqn{n_k} is the sample size of k-th cancer subtype.}

\item{lambda}{The common tuning parameter for controlling the overall degree sparsity of the estiamted gene networks. For details, refer to Supplementary TableS7.}

\item{tau}{The tuning parameter balances the network size between non-cancerous and cancerous networks. For details, refer to Supplementary TableS7.}

\item{delta}{The tuning parameter balances the network size between subtype specific networks. For details, refer to Supplementary TableS7.}
}
\value{
\item{\code{theta.y}}{A list (length = \eqn{K}) of estimated precision matrices from cancerous cells of different  cancer subtyes.}
\item{\code{theta.z}}{A matrix of estimated  precision matrice from non-cancerous cells shared by all cancer subtyes.}
\item{\code{LL.temp}}{Log-likelihood of the data for different EM iterations.}
}
\description{
The complete procedure for reconstructing multiple gene networks
 using NETI2. For details, refer to Supplementary Section S3.2.
}
\details{
The function is used to jointly reconstruction of multiple gene networks by simultaneously capturing inter-tumor
   and intra-tumor heterogeneity. For each cancer subtype, the observed gene  expression levels of tumor samples are assumed to be a
 mixture of expressions from non-cancerous and cancerous cells. The gene  expression levels of cancerous cells are assumed to follow subtype
 specific multivariate normal distributions, and the gene expression levels of non-cancerous cells across all subtypes are assumed to follow the
 same multivariate normal distribution. The precision matrices of these multivariate normal distributions are used to build the non-cancerous and subtype
 specific cancerous networks.  Given the observed gene expression data and tumor purity information data,
we use a penalized likelihood approach to estimate the model parameters. We develop an efficient iterative procedure based on Expectation
Maximization (EM) algorithm to solve the optimization problem with latent (unobserved) variables.
Each iteration of the EM algorithm consists of two steps: E-step and M-step. For details, refer to Supplementary Section S3.2.
}
\examples{
# Simulation data
data.x= generate.data(p = 100, n = 100, K = 4, network.type ="ER", umin = 0.5, umax = 1)
result = NETI2(data.x$X,data.x$purity, lambda = 0.9, tau = 0.5,delta = 0.5)

# TCGA breast cancer data
data("TCGA.BRCA")
result = NETI2(TCGA.BRCA$BRCA.data,TCGA.BRCA$BRCA.purity, lambda = 1.6, tau = 0.4,delta = 0.2)
}
\references{
Jia-Juan Tu, Le Ou-Yang, Hong Yan, Xiao-Fei Zhang and Hong Qin (2019), Joint reconstruction of multiple gene networks by simultaneously capturing inter-tumor
     and intra-tumor heterogeneity
}
\seealso{
{ \code{\link{generate.data}}, \code{\link{TCGA.BRCA}}}
}
\author{
Jia-Juan Tu
}
