% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA.BRCA}
\alias{TCGA.BRCA}
\title{TCGA breast cancer data and  tumor purity information data}
\format{An object of class \code{list} of length 2.}
\source{
[1] The Cancer Genome Atlas Research Network (2012), Comprehensive molecular portraits of human breast tumors. Nature. 490 (7418), 61-70. (http://cancergenome.nih.gov/)

[2] Aran, D. et al. (2015). Systematic pan-cancer analysis of tumour purity. Nature Communications, 6, 8971.
}
\usage{
data("TCGA.BRCA")
}
\description{
The TCGA  breast cancer gene expression data and tumor purity information data are used in our study.
The gene expression data are obtained from the TCGA database and collected  from mRNA expression
(Agilent G450 microarray). The breast cancer is a heterogeneous disease that can be mainly divided into four molecular
subtypes: Luminal A, Luminal B, Basal-like and HER2-enriched.
The data only include expression measurements of genes that overlap
with the breast cancer pathway collected from the Kyoto Encyclopedia of Genes and Genomes database.
The tumor purity information  of these samples are collected from Aran et al (2015).
This data is an object of class list of length two. The first element (BRCA.data) is the gene expression  data and the second element (BRCA.purity)
is the tumor purity information data.
}
\examples{
data("TCGA.BRCA")
}
\references{
Jia-Juan Tu, Le Ou-Yang, Hong Yan, Xiao-Fei Zhang and Hong Qin (2019), Joint reconstruction of multiple gene networks by simultaneously capturing inter-tumor
     and intra-tumor heterogeneity.
}
\seealso{
{\code{\link{NETI2}}, \code{\link{generate.data}}}
}
\author{
Jia-Juan Tu
}
\keyword{datasets}
